refactor(tools): move native tools out of NAT
