+++ DEPENDENCIES.md
@@ -119,23 +119,25 @@ xr-ai-models  (agent-sdk/xr-ai-models/)
     own httpx wrappers. Profiles may separate adapter, endpoint, and deployment
     metadata while the existing flat YAML schema remains valid.
 
-xr-ai-nat  (agent-sdk/xr-ai-nat/)
+xr-ai-tools  (agent-sdk/xr-ai-tools/)
     └── nemo-relay >=0.7.2,<0.8
     └── pydantic >=2.10
     └── [relay] xr-ai-models [editable: ../xr-ai-models]
     └── [live-vision] numpy >=1.24, Pillow >=10.0, xr-ai-hub-client [editable: ../xr-ai-hub-client], xr-ai-models [editable: ../xr-ai-models]
-    └── [agents] nvidia-nat-core ==1.8.0, nvidia-nat-langchain ==1.8.0, xr-ai-models [editable: ../xr-ai-models]
-    └── [mcp] nvidia-nat-core ==1.8.0, fastmcp >=3.4,<4
-    └── [services] nvidia-nat-core ==1.8.0, msgpack >=1.0, pyzmq >=27.0
-    └── [vision] nvidia-nat-core ==1.8.0, httpx >=0.27, numpy >=1.24, Pillow >=10.0, xr-ai-hub-client [editable: ../xr-ai-hub-client], xr-ai-models [editable: ../xr-ai-models]
-    └── [voice] nvidia-nat-core ==1.8.0, xr-ai-voice [editable: ../xr-ai-voice]
-    The base package is the toolkit-independent native tools layer: Pydantic
-    request and response models, Relay-managed execution, the generic
-    ``AgentRunner`` protocol, and a bounded default tool loop over
-    `xr-ai-models`. The ``[relay]`` and
-    ``[live-vision]`` extras add model-backed tools without selecting NeMo Agent
-    Toolkit. The existing function groups remain behind legacy compatibility
-    extras while they migrate. The ``xr_spatial_math`` function group accepts
+    Toolkit-independent native tools: Pydantic request and response models,
+    Relay-managed execution, the generic ``AgentRunner`` protocol, a bounded
+    default tool loop, and participant-scoped live vision.
+
+xr-ai-nat  (agent-sdk/xr-ai-nat/)
+    └── nvidia-nat-core ==1.8.0
+    └── pydantic >=2.10
+    └── [agents] nvidia-nat-langchain ==1.8.0, xr-ai-models [editable: ../xr-ai-models]
+    └── [mcp] fastmcp >=3.4,<4
+    └── [services] msgpack >=1.0, pyzmq >=27.0
+    └── [vision] httpx >=0.27, numpy >=1.24, Pillow >=10.0, xr-ai-hub-client [editable: ../xr-ai-hub-client], xr-ai-models [editable: ../xr-ai-models]
+    └── [voice] xr-ai-voice [editable: ../xr-ai-voice]
+    Typed, in-process NeMo Agent Toolkit functions retained while their
+    concrete capabilities migrate. The ``xr_spatial_math`` function group accepts
     explicit coordinate frames and
     performs deterministic spatial calculations without OpenXR, model, or MCP
     dependencies. ``xr_text_memory`` owns persistent per-source JSONL text
@@ -334,7 +336,8 @@ vec-mcp-server  (agent-mcp-servers/vec-mcp/)
 xr-ai-tests  (tests/)
     └── xr-ai-hub-client             [editable: ../agent-sdk/xr-ai-hub-client]
     └── xr-ai-models            [editable: ../agent-sdk/xr-ai-models]
-    └── xr-ai-nat[agents,relay,services,vision] [editable: ../agent-sdk/xr-ai-nat]
+    └── xr-ai-nat[agents,services,vision] [editable: ../agent-sdk/xr-ai-nat]
+    └── xr-ai-tools[relay,live-vision] [editable: ../agent-sdk/xr-ai-tools]
     └── xr-rag-service [editable: ../services/rag-service]
     └── xr-ai-pipecat           [editable: ../agent-sdk/xr-ai-pipecat]
     └── xr-ai-voice             [editable: ../agent-sdk/xr-ai-voice]
@@ -554,7 +557,7 @@ the latest video frame via streaming VLM and replies with both
 | Sub-project | Package | Internal deps | External deps |
 |---|---|---|---|
 | Orchestrator | `simple-vlm-example` | `xr-ai-launcher` | — |
-| Worker | `simple-vlm-example-worker` | `xr-ai-hub-client [editable]`, `xr-ai-logging [editable]`, `xr-ai-models [editable]`, `xr-ai-nat[relay,live-vision] [editable]`, `xr-ai-voice [editable]`, `xr-ai-voicegate [editable]` | loguru >=0.7, pyyaml >=6.0 (`xr-ai-voice` pulls in VAD, pipecat-ai, numpy, and scipy; `xr-ai-nat[live-vision]` pulls in numpy and Pillow) |
+| Worker | `simple-vlm-example-worker` | `xr-ai-hub-client [editable]`, `xr-ai-logging [editable]`, `xr-ai-models [editable]`, `xr-ai-tools[relay,live-vision] [editable]`, `xr-ai-voice [editable]`, `xr-ai-voicegate [editable]` | loguru >=0.7, pyyaml >=6.0 (`xr-ai-voice` pulls in VAD, pipecat-ai, numpy, and scipy; `xr-ai-tools[live-vision]` pulls in numpy and Pillow) |
 
 The packaged worker constructs the finite `LiveVisionTool` used by agentic
 flows, then maps its separate direct-voice `LiveVisionResponder` to
