+++ agent-samples/simple-vlm-example/worker/pyproject.toml
@@ -13,7 +13,7 @@ dependencies = [
     "xr-ai-hub-client",
     "xr-ai-logging",
     "xr-ai-models",
-    "xr-ai-nat[relay,live-vision]",
+    "xr-ai-tools[relay,live-vision]",
     "xr-ai-voice",
     "xr-ai-voicegate",
     "loguru>=0.7",
@@ -24,7 +24,7 @@ dependencies = [
 xr-ai-hub-client = { path = "../../../agent-sdk/xr-ai-hub-client", editable = true }
 xr-ai-logging   = { path = "../../../utils/xr-ai-logging", editable = true }
 xr-ai-models    = { path = "../../../agent-sdk/xr-ai-models", editable = true }
-xr-ai-nat       = { path = "../../../agent-sdk/xr-ai-nat", editable = true }
+xr-ai-tools     = { path = "../../../agent-sdk/xr-ai-tools", editable = true }
 xr-ai-voice     = { path = "../../../agent-sdk/xr-ai-voice", editable = true }
 xr-ai-voicegate = { path = "../../../utils/xr-ai-voicegate", editable = true }
