+++ agent-sdk/xr-ai-nat/xr_ai_nat/__init__.py
@@ -1,15 +1,6 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Toolkit-independent native XR tools with legacy NAT compatibility."""
+"""NAT-native XR functions."""
 
-from .agent_runner import AgentRunner, as_agent_tool
-from .tools import Tool, ToolInvocationResult, ToolSet
-
-__all__ = [
-    "AgentRunner",
-    "Tool",
-    "ToolInvocationResult",
-    "ToolSet",
-    "as_agent_tool",
-]
+__all__: list[str] = []
