+++ agent-sdk/xr-ai-tools/pyproject.toml
@@ -1,0 +1,27 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "xr-ai-tools"
+version = "0.1.0"
+description = "Native Relay-managed tools for XR AI."
+requires-python = ">=3.11,<3.13"
+dependencies = [
+    "nemo-relay>=0.7.2,<0.8",
+    "pydantic>=2.10",
+]
+
+[project.optional-dependencies]
+relay = ["xr-ai-models"]
+live-vision = ["numpy>=1.24", "Pillow>=10.0", "xr-ai-hub-client", "xr-ai-models"]
+
+[tool.uv.sources]
+xr-ai-models = { path = "../xr-ai-models", editable = true }
+xr-ai-hub-client = { path = "../xr-ai-hub-client", editable = true }
+
+[tool.hatch.build.targets.wheel]
+packages = ["xr_ai_tools"]
