+++ agent-sdk/xr-ai-tools/xr_ai_tools/__init__.py
@@ -1,0 +1,15 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Toolkit-independent native XR tools."""
+
+from .agent_runner import AgentRunner, as_agent_tool
+from .tools import Tool, ToolInvocationResult, ToolSet
+
+__all__ = [
+    "AgentRunner",
+    "Tool",
+    "ToolInvocationResult",
+    "ToolSet",
+    "as_agent_tool",
+]
