+++ agent-sdk/xr-ai-tools/xr_ai_tools/_pixels.py
