+++ tests/pyproject.toml
@@ -14,7 +14,8 @@ dependencies = [
     # Pulled in via editable installs of the workspace packages.
     "xr-ai-hub-client",
     "xr-ai-models",
-    "xr-ai-nat[agents,relay,services,vision]",
+    "xr-ai-nat[agents,services,vision]",
+    "xr-ai-tools[relay,live-vision]",
     "xr-ai-pipecat",
     "xr-ai-voice",
     "xr-media-hub",
@@ -46,6 +47,7 @@ dependencies = [
 xr-ai-hub-client           = { path = "../agent-sdk/xr-ai-hub-client",                        editable = true }
 xr-ai-models          = { path = "../agent-sdk/xr-ai-models",           editable = true }
 xr-ai-nat             = { path = "../agent-sdk/xr-ai-nat",              editable = true }
+xr-ai-tools           = { path = "../agent-sdk/xr-ai-tools",            editable = true }
 xr-ai-pipecat         = { path = "../agent-sdk/xr-ai-pipecat",          editable = true }
 xr-ai-voice           = { path = "../agent-sdk/xr-ai-voice",            editable = true }
 xr-media-hub          = { path = "../services/xr-media-hub",             editable = true }
